+++ lib/app_settings.dart
@@ -1,2 +1,3 @@
 export 'src/app_settings.dart';
 export 'src/app_settings_type.dart';
+export 'src/app_settings_panel_type.dart';