+++ lib/app_settings_method_channel.dart
@@ -8,13 +8,23 @@ import 'app_settings_platform_interface.dart';
 class MethodChannelAppSettings extends AppSettingsPlatform {
   /// The method channel used to interact with the native platform.
   @visibleForTesting
-  final methodChannel = const MethodChannel('app_settings');
+  final methodChannel = const MethodChannel('com.spencerccf.app_settings/methods');
 
   @override
   Future<void> openAppSettings({
     AppSettingsType type = AppSettingsType.settings,
     bool asAnotherTask = false,
   }) {
-    return methodChannel.invokeMethod<void>(type.name, {'asAnotherTask': asAnotherTask});
+    return methodChannel.invokeMethod<void>('openSettings', {
+      'asAnotherTask': asAnotherTask,
+      'type': type.name,
+    });
+  }
+
+  @override
+  Future<void> openAppSettingsPanel(AppSettingsPanelType type) {
+    return methodChannel.invokeMethod<void>('openSettingsPanel', {
+      'type': type.name,
+    });
   }
 }